Chatbot/Chatbot/PullingBots/requirements.txt
selenium==4.15.2
pandas==2.1.3
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread-dataframe==3.3.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
PyMuPDF==1.23.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncio==3.4.3
playwright==1.40.0
playwright-stealth
sqlalchemy[asyncio]
python-dotenv
asyncpg
openai
requests
beautifulsoup4
tqdm
pyap
pdf2image
pytesseract
opencv-python
# LangChain dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
langchain-core==0.1.12
# Additional dependencies for modular scraper
pydantic==2.5.2
click==8.1.7
jinja2==3.1.2
lxml==4.9.3
jsonschema==4.20.0
pyyaml==6.0.1
rich==13.7.0 